+++ City.h
@@ -11,16 +11,16 @@ class City
 {
 public:
 	City(string name, string coordinates);
-	City(string name, long double latitude, long double longitude);
+	City(string name, double latitude, double longitude);
 	~City();
 
 	void setName(string name);
-	void setLatitude(long double latitude);
-	void setLongitude(long double longitude);
+	void setLatitude(double latitude);
+	void setLongitude(double longitude);
 
 	string getName();
-	long double getLatitude();
-	long double getLongitude();
+	double getLatitude();
+	double getLongitude();
 
 	string getCoordinates();
 	
@@ -31,8 +31,8 @@ public:
 private:
 
 	string m_name; 
-	long double m_latitude;
-	long double m_longitude;
+	double m_latitude;
+	double m_longitude;
 	string m_coordinates;
 
 };
